chore: api 文件夹合并为单文件 api.py，修复 basedpyright 错误
