+++ typos.toml
@@ -2,3 +2,4 @@
 detecter = "detecter"
 Detecter = "Detecter"
 mapp = "mapp"
+pn =  "pn"
